Small rename/reorg of client package
